LangChain_Hands_on_Intro/requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pandas
langchain_ollama
streamlit